+++ lib/main.dart
@@ -1,20 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:hive/hive.dart';
-import 'package:hive_flutter/hive_flutter.dart';
 import 'package:provider/provider.dart';
 import 'package:shop/controller/mainScreenController.dart';
 import 'package:shop/controller/productController.dart';
-import 'package:shop/visuals/screens/primary_pages/MyHomePage.dart';
+import 'package:shop/visuals/screens/MyHomePage.dart';
 import 'package:shop/visuals/screens/mainScreen.dart';
-import 'package:path_provider/path_provider.dart' as path_provider;
 
-void main() async {
+void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  await Hive.initFlutter();
-  await Hive.openBox("fav_box");
-  await Hive.openBox("cart_box");
-
   SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp])
       .then((_) {
     runApp(MultiProvider(providers: [
