+++ lib/mainScreen.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
 import 'package:provider/provider.dart';
 import 'package:shop/controller/mainScreenController.dart';
-import 'package:shop/visuals/screens/primary_pages/MyHomePage.dart';
-import 'package:shop/visuals/screens/primary_pages/cart_page.dart';
-import 'package:shop/visuals/screens/primary_pages/profile.dart';
-import 'package:shop/visuals/screens/primary_pages/search_page.dart';
-import 'package:shop/visuals/shared/widgets/bottomBarButton.dart';
-import 'package:shop/visuals/shared/widgets/bottomNavigationBar.dart';
+import 'package:shop/visuals/screens/MyHomePage.dart';
+import 'package:shop/visuals/screens/cart_page.dart';
+import 'package:shop/visuals/screens/profile.dart';
+import 'package:shop/visuals/screens/search_page.dart';
+import 'package:shop/visuals/shared/bottomBarButton.dart';
+import 'package:shop/visuals/shared/bottomNavigationBar.dart';
 
 class mainScreen extends StatefulWidget {
   @override
@@ -29,7 +29,9 @@ class _mainScreenState extends State<mainScreen> {
     return Consumer<MainScreenNotifier>(
       builder: (context, mainScreenNotifier, child) {
         int currentIndex = mainScreenNotifier.pageIndex;
-
+        if (kDebugMode) {
+          print("current Page Index " + currentIndex.toString());
+        }
         return Scaffold(
           backgroundColor: const Color(0xFFE2E2E2),
           bottomNavigationBar: SafeArea(
