+++ lib/visuals/screens/cart_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class CartScreen extends StatelessWidget {
+  const CartScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Text("Cart Screen");
+  }
+}
