+++ lib/visuals/screens/search_page.dart
