+++ lib/visuals/shared/filterButton.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
-import 'package:shop/visuals/shared/utilities/appStyles.dart';
+import 'package:shop/visuals/shared/appStyles.dart';
 
 Future<dynamic> filter(BuildContext context) {
   double _minValue = 10000;
