+++ lib/visuals/shared/homeWidget.dart
@@ -1,12 +1,11 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:shop/controller/productController.dart';
+import 'package:flutter/widgets.dart';
 import 'package:shop/model/sneakerModel.dart';
-import 'package:shop/visuals/screens/secondary_pages/descriptionScreen.dart';
-import 'package:shop/visuals/screens/secondary_pages/productByCat.dart';
-import 'package:shop/visuals/shared/utilities/appStyles.dart';
-import 'package:shop/visuals/shared/widgets/productCard.dart';
+import 'package:shop/visuals/screens/descriptionScreen.dart';
+import 'package:shop/visuals/screens/productByCat.dart';
+import 'package:shop/visuals/shared/appStyles.dart';
+import 'package:shop/visuals/shared/productCard.dart';
 
 class HomeWidget extends StatefulWidget {
   final int TabIndex;
@@ -20,7 +19,6 @@ class HomeWidget extends StatefulWidget {
 class _HomeWidgetState extends State<HomeWidget> {
   @override
   Widget build(BuildContext context) {
-    var productNotifier = Provider.of<ProductNotifier>(context);
     return SizedBox(
       height: MediaQuery.of(context).size.height,
       child: SingleChildScrollView(
@@ -33,7 +31,7 @@ class _HomeWidgetState extends State<HomeWidget> {
                   child: FutureBuilder<List<Sneaker>>(
                     builder: (context, snapshot) {
                       if (snapshot.connectionState == ConnectionState.waiting) {
-                        return const CircularProgressIndicator.adaptive();
+                        return CircularProgressIndicator.adaptive();
                       } else if (snapshot.hasError) {
                         return Text(
                           snapshot.error.toString(),
@@ -49,23 +47,16 @@ class _HomeWidgetState extends State<HomeWidget> {
                               itemBuilder: (context, index) {
                                 final shoeData = shoelist[index];
                                 return GestureDetector(
-                                  onTap: () {
-                                    final sizes = shoeData.sizes;
-                                    productNotifier.SetShoeSize = sizes;
-                                    Navigator.of(context).push(
-                                      MaterialPageRoute(
-                                        builder: (context) {
-                                          return DescriptionScreen(
-                                            id: shoeData.id,
-                                            shoeCategory: shoeData.category,
-                                          );
-                                        },
+                                  onTap: () => Navigator.of(context).push(
+                                    MaterialPageRoute(
+                                      builder: (context) => DescriptionScreen(
+                                        id: shoeData.id,
+                                        shoeCategory: shoeData.category,
                                       ),
-                                    );
-                                  },
+                                    ),
+                                  ),
                                   child: Container(
-                                    margin:
-                                        const EdgeInsets.fromLTRB(2, 0, 0, 0),
+                                    margin: EdgeInsets.fromLTRB(2, 0, 0, 0),
                                     width:
                                         MediaQuery.of(context).size.width * 0.6,
                                     child: ProductCard(
@@ -119,7 +110,7 @@ class _HomeWidgetState extends State<HomeWidget> {
                     future: widget.future,
                     builder: (context, snapshot) {
                       if (snapshot.connectionState == ConnectionState.waiting) {
-                        return const CircularProgressIndicator.adaptive();
+                        return CircularProgressIndicator.adaptive();
                       } else if (snapshot.hasError) {
                         return Text(
                           snapshot.error.toString(),
@@ -137,20 +128,14 @@ class _HomeWidgetState extends State<HomeWidget> {
                             itemBuilder: (context, index) {
                               final shoeData = snapshot.data![index];
                               return GestureDetector(
-                                onTap: () {
-                                  // productNotifier.SetShoeSize = shoeData.sizes;
-
-                                  Navigator.of(context).push(
-                                    MaterialPageRoute(
-                                      builder: (context) {
-                                        return DescriptionScreen(
-                                          id: shoeData.id,
-                                          shoeCategory: shoeData.category,
-                                        );
-                                      },
+                                onTap: () => Navigator.of(context).push(
+                                  MaterialPageRoute(
+                                    builder: (context) => DescriptionScreen(
+                                      id: shoeData.id,
+                                      shoeCategory: shoeData.category,
                                     ),
-                                  );
-                                },
+                                  ),
+                                ),
                                 child: Container(
                                   padding: const EdgeInsets.all(2),
                                   margin:
