+++ pubspec.yaml
@@ -46,9 +46,6 @@ dependencies:
   hive_storage: ^0.0.3
   gap: ^3.0.1
   staggered_grid_view_flutter: ^0.0.4
-  hive_flutter: ^1.1.0
-  hive_generator: ^2.0.1
-  toast: ^0.3.0
   
 
 dev_dependencies:
